legal-workbench/_archived/hub/requirements.txt
python-fasthtml>=0.12.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
httpx>=0.27.0
